move commands to management cog, more professional startup sequence, add file logger, separate out signup into a new cog
