+++ .gitignore
@@ -1,3 +1,4 @@
 appsettings.local*
 src/__pycache__
 *.pyc
+discord.log
